third_party/blink/web_tests/platform/linux/external/wpt/selection/contenteditable/initial-selection-on-focus.tentative_div-expected.txt

This is a testharness.js-based test.
Found 4 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] editor should collapse selection before the non-editable <p> node having a text node and followed by another text node
  assert_equals: Only one caret should be in the editor expected 1 but got 0
[FAIL] editor should collapse selection before the non-editable <p> node having a <br> node and followed by a text node
  assert_equals: Only one caret should be in the editor expected 1 but got 0
[FAIL] editor should collapse selection before the non-editable <p> node having text node and followed by a <br> node
  assert_equals: Only one caret should be in the editor expected 1 but got 0
[FAIL] editor should collapse selection before the non-editable <p> node having a <br> node and followed by another <br> node
  assert_equals: Only one caret should be in the editor expected 1 but got 0
Harness: the test ran to completion.
